aws-sagemaker-pre-built-algorithm/requirements.txt
boto
sagemaker